Avoid race conditions in tests
Splunk client uses an asynchronous HTTP queue, so we need to wait for mockserver to receive the call before trying to assert.
Create an abstract class for the mock-server lifecyle.
