+++ deployment/src/test/java/io/quarkiverse/logging/splunk/LoggingSplunkTest.java
@@ -7,7 +7,6 @@ package io.quarkiverse.logging.splunk;
 import static org.hamcrest.MatcherAssert.*;
 import static org.hamcrest.Matchers.*;
 import static org.mockserver.model.HttpRequest.request;
-import static org.mockserver.model.HttpResponse.response;
 import static org.mockserver.model.JsonBody.json;
 import static org.mockserver.model.RegexBody.regex;
 
@@ -19,19 +18,14 @@ import org.jboss.logging.MDC;
 import org.jboss.logmanager.handlers.DelayedHandler;
 import org.jboss.shrinkwrap.api.ShrinkWrap;
 import org.jboss.shrinkwrap.api.spec.JavaArchive;
-import org.junit.jupiter.api.AfterAll;
-import org.junit.jupiter.api.AfterEach;
-import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.RegisterExtension;
-import org.mockserver.integration.ClientAndServer;
-import org.mockserver.model.ClearType;
 import org.mockserver.verify.VerificationTimes;
 
 import io.quarkus.bootstrap.logging.InitialConfigurator;
 import io.quarkus.test.QuarkusUnitTest;
 
-class LoggingSplunkTest {
+class LoggingSplunkTest extends AbstractMockServerTest {
 
     @RegisterExtension
     static final QuarkusUnitTest unitTest = new QuarkusUnitTest()
@@ -39,29 +33,7 @@ class LoggingSplunkTest {
             .withConfigurationResource("mock-server.properties")
             .setArchiveProducer(() -> ShrinkWrap.create(JavaArchive.class));
 
-    static ClientAndServer httpServer;
-
     static final org.jboss.logging.Logger logger = org.jboss.logging.Logger.getLogger(LoggingSplunkTest.class);
-
-    @BeforeAll
-    public static void setUpOnce() {
-        httpServer = ClientAndServer.startClientAndServer(8088);
-        // This needs to be done as early as possible, so Quarkus startup logs don't fail to be sent
-        httpServer
-                .when(request().withPath("/services/collector/event/1.0"))
-                .respond(response().withStatusCode(200).withBody("{}"));
-        httpServer.when(request()).respond(response().withStatusCode(400));
-    }
-
-    @AfterAll
-    public static void tearDownOnce() {
-        httpServer.stop();
-    }
-
-    @AfterEach
-    public void tearDown() {
-        httpServer.clear(request(), ClearType.LOG);
-    }
 
     @Test
     void handlerShouldBeCreated() {
@@ -76,18 +48,21 @@ class LoggingSplunkTest {
     @Test
     void handlerShouldFormatMessage() {
         logger.warnv("hello {0}", "splunk!");
+        awaitMockServer();
         httpServer.verify(request().withBody(json("{ event: { message: 'hello splunk!' }}")));
     }
 
     @Test
     void eventIsAJsonObjectWithMetadata() {
         logger.warn("hello splunk");
+        awaitMockServer();
         httpServer.verify(request().withBody(json("{ event: { message: 'hello splunk' }}")));
     }
 
     @Test
     void eventHasStandardMetadata() {
         logger.warn("hello splunk");
+        awaitMockServer();
         httpServer.verify(request().withBody(json(
                 "{ source: 'mysource', sourcetype: 'mysourcetype', index: 'myindex'} "))
                 .withBody(regex(".*host.*")));
@@ -96,12 +71,14 @@ class LoggingSplunkTest {
     @Test
     void tokenIsSentAsAuthorizationHeader() {
         logger.warn("hello splunk");
+        awaitMockServer();
         httpServer.verify(request().withHeader("Authorization", "Splunk 12345678-1234-1234-1234-1234567890AB"));
     }
 
     @Test
     void clientAddsMinimalMetadata() {
         logger.warn("hello splunk");
+        awaitMockServer();
         httpServer.verify(request().withBody(json(
                 "{ event: { message: 'hello splunk', severity:'WARN' }}")));
     }
@@ -110,6 +87,7 @@ class LoggingSplunkTest {
     void mdcFieldsShouldBeSentAsMetadata() {
         MDC.put("mdc-key", "mdc-value");
         logger.warn("hello mdc");
+        awaitMockServer();
         httpServer.verify(
                 request().withBody(json("{ event: { message: 'hello mdc', properties: { 'mdc-key': 'mdc-value' }}}")));
     }
@@ -117,6 +95,7 @@ class LoggingSplunkTest {
     @Test
     void messageShouldContainException() {
         logger.error("unexpected error", new RuntimeException("test exception"));
+        awaitMockServer();
         httpServer.verify(request()
                 .withBody(regex(".*unexpected error: java.lang.RuntimeException: test exception.*")));
     }
